[Unity] Extend RemoveAllUnused to support relax::Expr (#15807)
* [Unity] Extend RemoveAllUnused to support relax::Expr

Prior to this commit, the `RemoveAllUnused` utility could only
apply to `relax::Function` instances.  This commit extends the allowed
usage to apply to any `relax::Expr` that contains variable bindings,
such as `relax::SeqExpr`.

* Update unit tests for new behavior

Because `RemoveAllUnused` now handles unused variables in a
non-dataflow binding block, passes that use it as a utility function
needed a few tests updated.