[CI] Update ci-gpu image (#15836)
* [CI] Update ci-gpu image

The ci-gpu is updated to make changes in #15819 reflect in CI.

* update image that contains oneflow==0.8.0

* Force old coreml api to fix failing test

Between the previous ci_gpu image and the one in use now, the version of
coreml has changed since it is unpinned (6.3.0 -> 7.0.0). This caused a
test failure in the coreml frontend tests. The problem is that coreml
now uses a new default format for models. This commit forces coreml to
use the old model format so the test passes.

More information can be found here: https://apple.github.io/coremltools/docs-guides/source/convert-learning-models.html