[Unity][WebGPU] Allow lower max storage buffer binding size (#16108)
This PR lowers the required `maxStorageBufferBindingSize` from 1GB to 128MB when necessary.

Previously, we required 1GB, which led to errors when running on some devices (e.g. Android's Chrome, since Android has 128MB as its limit).

Now, if the device does not allow 1GB, we lower it to 128MB. We only throw an error when 128MB still exceeds the limit.