[Unity] Check usage location when canonicalizing trivial bindings (#16193)
* [Unity] Check usage location when canonicalizing trivial bindings

Prior to this commit, the `CanonicalizeBindings` transform handled
trivial bindings, including binding of a `DataflowVar` to a `Var`, and
usage of `Var` where a `DataflowVar` should be used instead.  However,
it did not handle the case where both of these occur at the same
time.  That is, a binding of `var_x = dataflow_y`, where `var_x` is
only used within the dataflow block.  For these cases,
`CanonicalizeBindings` would need to be run repeatedly.

This commit updates `CanonicalizeBindings` to check if a trivial
binding is used outside the dataflow block.  This avoids producing
output that requires further canonicalization.

* Add unit test following review comments