[Unity][Transform] Update LambdaLift to use name of lifted lambda (#16306)
* [Unity][Transform] Update LambdaLift to use name of lifted lambda

Prior to this commit, the `LambdaLift` pass named each function
as `"lifted_func_" + i`, in incremental order of occurrence.  This
provided unique names for each function, but could be difficult to
read, or to refer to the lifted functions.  This commit updates the
naming scheme to use the location at which the lifted lambda occurs to
generate a unique name for the new `GlobalVar`.

* Update variables names and comments for unique function naming

* Add unit test for conflicting name