[Metal] Dispatch numerically stable tanh for metal (#16438)
Prior to this PR, `tanh(x)`returns `NaN` on metal when `x > 45.0`.

Metal's built-in tanh is implemented as `(t - 1.0) / (t + 1.0)`, where `t = exp(2.0 * x)`. Hence for large `x`, `t` becomes `inf`, causing `tanh(x)` to be `NaN`.

A numerically stable `tanh` is implemented for `llvm`, this PR lifts it to `src/target/intrin_rule.cc` and apply the same rule for metal as well.