[Disco] Reduce Process/ThreadSession message queue reads and writes (#16817)
This PR reduces the number of reads and writes for the message queue
of ProcessSession and ThreadSession in Disco by caching all the data
to read/write.

The message queue in ThreadSession prior to this PR grabs the mutex
for multiple times for a batch of data to read/write. This PR enables
to read/write data from/to a local buffer first, and then read/write
from/to the critical region together. This reduces the number of
grabbing mutex to once.

The message queue in ProcessSession prior to this PR reads/writes
the inter-process pipe for multiple times for a batch of data.
This PR uses a local buffer to cache all the data first, and then
issues a single read/write from/to the pipe, and effectively reduces
the number of reads/writes to the pipe, which may causes extra
system overhead.