[Relax] Enhance symbolic expr estimation in memory planning (#16872)
This PR enhances the symbolic expression upper bound estimation in
static memory planning.

Prior to this PR, we are not able to estimate the upper bound of
`a * b` when `a` has an upper bound while `b` does not. This PR
enhances the estimation with arith::IntSet.

We introduce another TIR attribute `tir_non_negative_var` to indicate
the non-negative TIR variables for memory planning use.

A new unit test is introduced for this enhancement.