[QoL][Relax] Return well-formed IR from relax::Function::CreateEmpty (#16861)
Prior to this commit, the static method `relax::Function::CreateEmpty`
returned a function with a nullptr as the body.  While only intended for use in
bookkeeping for TVMScript, allowing nullptr in this location can cause
unexpected segfaults while debugging.  For example, adding a print
statement

This commit updates the `relax::Function::CreateEmpty` function to
contain a placeholder body, consistent with the `ret_struct_info`
argument provided.