[Web] Support string[] in setPackedFunc() and exceptionally long arrays (#16910)
There are two changes in this PR.

#### Change 1: Support `string[]` in `setPackedFunc()`

Prior to this PR, we cannot pass in `string[]` from typescript to
a TVM PackedFunc and need to convert it to `TVMArray<TVMString>`
(for instance in `getParamsFromCacheByName()`). This may not be the
most convenient if the PackedFunc's caller is not internal to tvmjs.
Thus, this PR moves the conversion to `setPackedFunc()` instead.

#### Change 2: Support exceptionally long TVM arrays

The second change is dealing with exceptionally long TVM arrays.
In cases like passing in a token table, we need to pass in a long
`string[]` (in Llama-3's case, of size 128000), leading to JS error
`RangeError: Maximum call stack size exceeded` since we treat each
string element as an argument, shown in `this.ctx.arrayMake(...inputs)`. 

This PR sets an empirical call stack limit of 30000 and chunks the
array elements in `makeTVMArray()`, converting each chunk to its
own TVMArray. Then we concatenate them with the newly implemented
`runtime.ArrayConcat` that concatenates N TVMArrays.

Tested end-to-end in WebLLM.