[SVE] Add codegen support for `vscale_range()` function attribute (#16962)
This commit adds support for the `vscale_range()` LLVM function attribute to be generated for SVE and SME targets.
Some LLVM optimisation passes make use of the `vscale_range()` function attribute when scalable vectors are present (e.g. BasicAA llvm/llvm-project/pull/80445), so we include it alongside the "target_cpu" and "target-features" attributes.