[CI] Reduce logging level when checking if docker image exists (#17221)
Prior to this commit, the `image_exists` utility in
`determine_docker_images.py` logged the full response for success, and
the full HTTP error if an exception is caught.  However, this is the
expected behavior when loading a docker image from `tlcpackstaging`,
such as the current images tagged with `20240428-060115-4cff01a53`.
Logging this fallback as an error makes it difficult to find the first
actual error that occurred in CI.

This commit updates these logging statments `logging.info` and
`logging.exception` to instead use `logging.debug`.