[Relax][Bugfix] FCallPacked not checked in CodegenVMTIR (#17073)
Prior to this commit, an operator's `FCallPacked` attribute, used to
specify a 1:1 mapping between a relax operator and a `PackedFunc` that
implements it, was only checked in `CodegenVM`.  Any operator with
`FCallPacked` would raise an error when compiled using `CodegenVMTIR`.

This commit removes the `FCallPacked` handling from `CodegenVM`
altogether, and instead checks for this attribute as part of
`LegalizeOps`.  This provides the same functionality across both
backends.