[Relax] Validate StructInfo of variable bindings (#17332)
* [Relax] Validate StructInfo of variable bindings

In Relax, both the variable and the expression in a `VarBinding` may
contain `StructInfo` annotations.  Prior to this commit, these
`StructInfo` annotations could be inconsistent, assigning an
expression to a variable of incompatible type.

This commit updates the Relax well-formed checker to verify that the
`StructInfo` of Relax variables accurately describes their contents.

* Fix unit tests

* [Relax][Bugfix] LCA of PrimStructInfo must check known values

The `StructInfoLCA` determines the lowest common ancestor between two
`StructInfo` annotations.  This is primarily used in Relax to
determine the appropriate `StructInfo` annotation for a `relax::If`
node, given the `StructInfo` of each branch.  Prior to this commit,
when determining the LCA of two `PrimStructInfo` annotations, the
`StructInfoLCA` function only inspected the datatype of
`PrimStructInfo` annotations, and did not check for known values.  For
example, the LCA of `R.Prim(value=T.int64(128))` and
`R.Prim(value=T.int64(64))` is `R.Prim("int64")`, but was incorrectly
determined as `R.Prim(value=T.int64(128))` by the `StructInfoLCA`
function.

This commit updates `StructInfoLCA` to inspect the known values of a
`PrimStructInfo`, as well as the datatype.