[TIR] Add `is_vector` Method to DataType class and update usages across Codebase (#17443)
* Refactor data_type.h and c_runtime_api.h

This commit refactors the `data_type.h` and `c_runtime_api.h` files. It introduces a new function `is_vector()` in the `DataType` class to check if a type is a vector type. Additionally, it adds a new constant `kTVMGridConstant` in the `TVMTypeCode` enum in `c_runtime_api.h`. These changes improve the code organization and provide better support for vector types.

* revert kTVMGridConstant

* lint fix